Bit more on state managers
I learnt how to access vector elements. end and begin are vector methods to get iterators. end returns the position AFTER the last element, but begin is THE first element.
vector methods front and back give the value of the first and last element. You can't do a '++' to increment through, as all you have is a value. '"hello"++' doesn't make any sense
I'm now up to "Being able to change the current state is of paramount "